1. 메시지 ENUM 처리 2. 애매한 에러 처리 수정 3. 코드 가독성 고려 4. 유효성 검사 5. DTO 도입 6. 예외처리 고도화

+++ src/main/java/com/example/Delivery/Review/SpringDataJPAReviewRepository.java
@@ -11,7 +11,7 @@ public interface SpringDataJPAReviewRepository extends JpaRepository<Review, Int
     List<Review> findByStoreId(long storeId);
 
     // 리뷰 작성 메소드
-    default Review writeReview(int memberId, int storeId, double rating, String content) {
+    default Review writeReview(long memberId,long storeId, double rating, String content) {
         Review review = new Review();
         review.setMemberId(memberId);
         review.setStoreId(storeId);
